remove hardcoded and mv
