+++ src/app.rs
@@ -10,7 +10,7 @@ use leptos_meta::*;
 use leptos_query::*;
 use leptos_router::*;
 
-use crate::leptos_extra::query;
+use crate::leptos_extra::query::{self, ServerQueryInput};
 use crate::leptos_extra::signal::{provide_signal, use_signal};
 use crate::nix::flake::GetNixFlake;
 use crate::nix::health::traits::Check;
@@ -113,64 +113,22 @@ fn Dashboard(cx: Scope) -> impl IntoView {
     }
 }
 
-/// An input element tied to a [leptos_use::Query]
-///
-/// TODO: remove hardcoded stuff
-#[component]
-fn QueryInput<S>(
-    cx: Scope,
-    #[allow(unused_variables)] serverfn: std::marker::PhantomData<S>,
-) -> impl IntoView
-where
-    S: ToString + FromStr + Hash + Eq + Clone + leptos::server_fn::ServerFn<leptos::Scope>,
-    <S as std::str::FromStr>::Err: Display,
-    ServerQueryVal<<S as leptos::server_fn::ServerFn<leptos::Scope>>::Output>:
-        Clone + Serializable + 'static,
-{
-    let (query, set_query) = use_signal::<S>(cx);
-    let (input_err, set_input_err) = create_signal(cx, None::<String>);
-    view! { cx,
-        <label for="flake-url">
-            "Load a Nix flake"
-            <input
-                list="some-flakes"
-                id="flake-url"
-                type="text"
-                class="w-full p-1 font-mono"
-                on:change=move |ev| {
-                    match FromStr::from_str(&event_target_value(&ev)) {
-                        Ok(url) => {
-                            set_query(url);
-                            set_input_err(None)
-                        }
-                        Err(e) => set_input_err(Some(e.to_string())),
-                    }
-                }
-
-                prop:value=move || query().to_string()
-            /> <span class="text-red-500">{input_err}</span>
-            // TODO: use local storage, and cache user's inputs
-            <datalist id="some-flakes">
-                <option value="github:nammayatri/nammayatri"></option>
-                <option value="github:srid/haskell-template"></option>
-                <option value="github:juspay/nix-browser"></option>
-                <option value="github:nixos/nixpkgs"></option>
-            </datalist>
-        </label>
-    }
-}
-
 /// Nix flake dashboard
 #[component]
 fn NixFlake(cx: Scope) -> impl IntoView {
     let title = "Nix Flake";
-    // TODO: Can we introduce struct, ServerQuery?
+    let suggestions: Vec<GetNixFlake> = vec![
+        "github:nammayatri/nammayatri".into(),
+        "github:srid/haskell-template".into(),
+        "github:juspay/nix-browser".into(),
+        "github:nixos/nixpkgs".into(),
+    ];
     let (flake_url, _) = use_signal::<GetNixFlake>(cx);
     let res = query::use_server_query::<GetNixFlake>(cx, flake_url);
     view! { cx,
         <Title text=title/>
         <h1 class="text-5xl font-bold">{title}</h1>
-        <QueryInput serverfn=(PhantomData::<GetNixFlake>)/>
+        <ServerQueryInput serverfn=(PhantomData::<GetNixFlake>) suggestions=suggestions/>
         <RefetchQueryButton res=res.clone() k=()/>
         <div class="my-1 text-left">
             // <SuspenseWithErrorHandling>{res.data}</SuspenseWithErrorHandling>
